Added strategy parameter to scanToken function
