+++ pokepaylib/src/main/java/jp/pokepay/pokepaylib/BankAPI/Transaction/CreateTransactionWithBill.java
@@ -7,44 +7,26 @@ import java.util.Map;
 
 import jp.pokepay.pokepaylib.BankAPI.BankRequest;
 import jp.pokepay.pokepaylib.BankAPI.BankRequestError;
+import jp.pokepay.pokepaylib.Parameters.TransactionStrategy;
 import jp.pokepay.pokepaylib.ProcessingError;
 import jp.pokepay.pokepaylib.Request;
 import jp.pokepay.pokepaylib.Responses.UserTransaction;
 
 public class CreateTransactionWithBill extends BankRequest {
 
-    public enum STRATEGY {
-        POINT_PREFERRED("point_preferred"),
-        MONEY_ONLY("money_only");
-
-        private final String strategy;
-
-        STRATEGY(String strategy) {
-            this.strategy = strategy;
-        }
-
-        public String getStrategy() {
-            return strategy;
-        }
-    }
-
     @NonNull
     public String billId;
     public String accountId;
     public Double amount;
     public String couponId;
-    public STRATEGY strategy;
+    public TransactionStrategy strategy;
 
-    public CreateTransactionWithBill(String billId, String accountId, Double amount,String couponId, STRATEGY strategy) {
+    public CreateTransactionWithBill(String billId, String accountId, Double amount,String couponId, TransactionStrategy strategy) {
         this.billId    = billId;
         this.accountId = accountId;
         this.amount    = amount;
         this.couponId = couponId;
         this.strategy = strategy;
-    }
-
-    public CreateTransactionWithBill(String billId, String accountId, Double amount,String couponId) {
-        this(billId, accountId, amount, couponId, STRATEGY.POINT_PREFERRED);
     }
 
     protected final String path() {
