+++ pokepaylib/src/main/java/jp/pokepay/pokepaylib/BankAPI/Transaction/CreateTransactionWithJwt.java
@@ -7,42 +7,23 @@ import java.util.Map;
 
 import jp.pokepay.pokepaylib.BankAPI.BankRequest;
 import jp.pokepay.pokepaylib.BankAPI.BankRequestError;
+import jp.pokepay.pokepaylib.Parameters.TransactionStrategy;
 import jp.pokepay.pokepaylib.ProcessingError;
 import jp.pokepay.pokepaylib.Request;
 import jp.pokepay.pokepaylib.Responses.JwtResult;
 
 public class CreateTransactionWithJwt extends BankRequest {
 
-    public enum STRATEGY {
-        POINT_PREFERRED("point_preferred"),
-        MONEY_ONLY("money_only");
-
-        private final String strategy;
-
-        STRATEGY(String strategy) {
-            this.strategy = strategy;
-        }
-
-        public String getStrategy() {
-            return strategy;
-        }
-    }
-
-
     @NonNull
     public String data;
     public String accountId;
     public String couponId;
-    public STRATEGY strategy;
+    public TransactionStrategy strategy;
 
-    public CreateTransactionWithJwt(String data, String accountId, String couponId, STRATEGY strategy) {
+    public CreateTransactionWithJwt(String data, String accountId, String couponId, TransactionStrategy strategy) {
         this.data = data;
         this.accountId = accountId;
         this.couponId = couponId;
-    }
-
-    public CreateTransactionWithJwt(String data, String accountId, String couponId) {
-        this(data, accountId, couponId, STRATEGY.POINT_PREFERRED);
     }
 
     protected final String path() {
