+++ pokepaylib/src/main/java/jp/pokepay/pokepaylib/Parameters/TransactionStrategy.java
@@ -1,0 +1,16 @@
+package jp.pokepay.pokepaylib.Parameters;
+
+public enum TransactionStrategy {
+    POINT_PREFERRED("point_preferred"),
+    MONEY_ONLY("money_only");
+
+    private final String strategy;
+
+    TransactionStrategy(String strategy) {
+        this.strategy = strategy;
+    }
+
+    public String getStrategy() {
+        return strategy;
+    }
+}
