+++ pokepaylib/src/main/java/jp/pokepay/pokepaylib/Pokepay.java
@@ -30,6 +30,7 @@ import jp.pokepay.pokepaylib.BankAPI.User.GetUserSettingUrl;
 import jp.pokepay.pokepaylib.OAuthAPI.OAuthRequestError;
 import jp.pokepay.pokepaylib.OAuthAPI.Token.ExchangeAuthCode;
 import jp.pokepay.pokepaylib.Parameters.Product;
+import jp.pokepay.pokepaylib.Parameters.TransactionStrategy;
 import jp.pokepay.pokepaylib.Pokeregi.BLEController;
 import jp.pokepay.pokepaylib.Responses.AccessToken;
 import jp.pokepay.pokepaylib.Responses.BankError;
@@ -149,18 +150,18 @@ public class Pokepay {
         }
 
         public UserTransaction scanToken(String token) throws ProcessingError, BankRequestError {
-            return scanToken(token, null, null, null,null);
+            return scanToken(token, null, null, null,null, TransactionStrategy.POINT_PREFERRED);
         }
 
-        public UserTransaction scanToken(String token, Double amount, String accountId, Product[] products, String couponId) throws ProcessingError, BankRequestError {
+        public UserTransaction scanToken(String token, Double amount, String accountId, Product[] products, String couponId, TransactionStrategy strategy) throws ProcessingError, BankRequestError {
             if (token.startsWith(getWwwBaseUrl() + "/cashtrays/")) {
                 final String uuid = token.substring((getWwwBaseUrl() + "/cashtrays/").length());
-                final CreateTransactionWithCashtray createTransactionWithCashtray = new CreateTransactionWithCashtray(uuid, accountId, couponId);
+                final CreateTransactionWithCashtray createTransactionWithCashtray = new CreateTransactionWithCashtray(uuid, accountId, couponId, strategy);
                 return createTransactionWithCashtray.send(accessToken);
             }
             else if (token.startsWith(getWwwBaseUrl() + "/bills/")) {
                 final String uuid = token.substring((getWwwBaseUrl() + "/bills/").length());
-                final CreateTransactionWithBill createTransactionWithBill = new CreateTransactionWithBill(uuid, accountId, amount,couponId);
+                final CreateTransactionWithBill createTransactionWithBill = new CreateTransactionWithBill(uuid, accountId, amount,couponId, strategy);
                 return createTransactionWithBill.send(accessToken);
             }
             else if (token.startsWith(getWwwBaseUrl() + "/checks/")) {
@@ -175,13 +176,13 @@ public class Pokepay {
             else {
                 String key = parseAsPokeregiToken(token);
                 if (key.length() > 0) {
-                    return scanTokenBLE(key, couponId);
+                    return scanTokenBLE(key, couponId, strategy);
                 }
             }
             throw new ProcessingError("Unknown token format");
         }
 
-        private UserTransaction scanTokenBLE(String token, String couponId) throws ProcessingError, BankRequestError {
+        private UserTransaction scanTokenBLE(String token, String couponId, TransactionStrategy strategy) throws ProcessingError, BankRequestError {
             if (context == null) {
                 throw new ProcessingError("Scanning to pokeregi requires Context (for BLE)");
             }
@@ -190,7 +191,7 @@ public class Pokepay {
                 bleController = new BLEController(token, context);
                 bleController.connect(1000 * 10);
                 final String jwt = bleController.read(1000 * 10);
-                final CreateTransactionWithJwt createTransactionWithJwt = new CreateTransactionWithJwt(jwt, null, couponId);
+                final CreateTransactionWithJwt createTransactionWithJwt = new CreateTransactionWithJwt(jwt, null, couponId, strategy);
                 final JwtResult jwtResult = createTransactionWithJwt.send(accessToken);
                 if (jwtResult.data != null) {
                     bleController.write(jwtResult.data, 1000 * 10);
